Processed feedback. Moved tests to separate file.
