+++ InfoLogger/test/lib/services/mocha-query-service.test.js
@@ -17,7 +17,6 @@ const sinon = require('sinon');
 const config = require('../../../config-default.js');
 const { QueryService } = require('../../../lib/services/QueryService.js');
 const { UnauthorizedAccessError, TimeoutError } = require('@aliceo2/web-ui');
-const { processPreparedSQLStatement } = require('../../../lib/utils/preparedStatementParser.js');
 
 describe('\'QueryService\' test suite', () => {
   const filters = {
@@ -204,39 +203,6 @@ describe('\'QueryService\' test suite', () => {
         sqlDataSource.queryFromFilters(realFilters, { limit: 10 }),
         new UnauthorizedAccessError('SQL: [ER_ACCESS_DENIED_ERROR, 1045] Access denied'),
       );
-    });
-
-    it('should be able to fill in a prepared statement and log it when executing a query', async () => {
-      const sqlDataSource = new QueryService(config.mysql);
-      sqlDataSource._logger = {
-        debugMessage: sinon.stub(),
-      };
-      sqlDataSource._pool = {
-        query: sinon.stub().resolves([{ hostname: 'test', severity: 'W' }]),
-      };
-      await sqlDataSource.queryFromFilters(realFilters, { limit: 10 });
-      const completeSqlQuery = "SELECT * FROM `messages` WHERE `timestamp`>='1563794601.351' AND" +
-        " `timestamp`<='1563794661.354' AND `hostname` = 'test' AND NOT(`hostname` = 'testEx' AND" +
-        " `hostname` IS NOT NULL) AND `severity` IN ('D','W') ORDER BY `TIMESTAMP` LIMIT 10;";
-      assert.strictEqual(sqlDataSource._logger.debugMessage.calledWith(`SQL to execute: ${completeSqlQuery}`), true);
-      // Test the processPreparedSQLStatement() function individually.
-      const requestedRows = 'SELECT * FROM `messages` WHERE `timestamp`>=? AND `timestamp`<=? AND `hostname` = ? '
-        + 'AND NOT(`hostname` = ? AND `hostname` IS NOT NULL) AND `severity` IN (?) ORDER BY `TIMESTAMP` LIMIT 10';
-      const values = [
-        '1563794601.351',
-        '1563794661.354',
-        'test',
-        'testEx',
-        [
-          'D',
-          'W',
-        ],
-      ];
-      const sqlProcessedResult = processPreparedSQLStatement(requestedRows, values, 10);
-      const expectedSqlResult = "SELECT * FROM `messages` WHERE `timestamp`>='1563794601.351' AND `timestamp`" +
-        "<='1563794661.354' AND `hostname` = 'test' AND NOT(`hostname` = 'testEx' AND `hostname` IS NOT" +
-        " NULL) AND `severity` IN ('D','W') ORDER BY `TIMESTAMP` LIMIT 10";
-      assert.strictEqual(sqlProcessedResult, expectedSqlResult);
     });
 
     it('should successfully return result when filters are provided for querying', async () => {
